Allergen and Nutrition up to date

+++ src/main/java/com/chefApp/demo/rest/AllergenEndpoint.java
@@ -1,6 +1,7 @@
 package com.chefApp.demo.rest;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -10,11 +11,9 @@ import org.springframework.web.bind.annotation.*;
 import com.chefApp.demo.controller.AllergenService;
 import com.chefApp.demo.model.Allergen;
 
-
 @RestController
 @RequestMapping("api/allergen")
 public class AllergenEndpoint {
-	
 	@Autowired
 	private AllergenService service;
 	
@@ -31,22 +30,36 @@ public class AllergenEndpoint {
 	}
 				
 	@PostMapping
-	public ResponseEntity<Allergen> createNewAllergen(@RequestBody Allergen allergen){
-		
-		Allergen saved = this.service.CreateOne(allergen);
-		
-		return new ResponseEntity<>(saved, HttpStatus.CREATED);
-	}
-	@PutMapping("{id}")
-	public Allergen updateById(@PathVariable long id, @RequestBody Allergen input) {	
-		return null;
+	public ResponseEntity<Allergen> createNewAllergen(@RequestBody Allergen allergen){	
+		return new ResponseEntity<>(this.service.CreateOne(allergen), HttpStatus.CREATED);
 	}
 	
+	@PutMapping("{id}")
+	public ResponseEntity<Allergen> updateById(@PathVariable() long id, @RequestBody Allergen input){
+		Allergen newAllergen = input;
+		Optional<Allergen> oldAllergen = this.service.getOne(input.getId());
+		if(oldAllergen.isEmpty()==false) {
+			Allergen updated = (Allergen) this.service.updateOne(newAllergen, oldAllergen.get().getId());
+			return new ResponseEntity<>(updated,HttpStatus.CREATED);
+		} else {
+			return new ResponseEntity<>(new Allergen(), HttpStatus.CONFLICT);
+		}
+	}
 	
 	@DeleteMapping("{id}")
-	public void  deleteAllergenById(@PathVariable() long id) {
-		service.deleteOne(id);
-		
-	}
-	
+	public ResponseEntity deleteAllergenById(@PathVariable()long id) {
+		if (id >= 0) {
+			Optional<Allergen> exists = service.getOne(id);
+			if(exists.isPresent()) {
+			service.deleteOne(id);
+			return new ResponseEntity(HttpStatus.ACCEPTED);
+		} else {
+			return new ResponseEntity(HttpStatus.NO_CONTENT);
+		} 
+		} else {
+			return new ResponseEntity(HttpStatus.BAD_REQUEST);
+		}
+	}		
 }
+	
+
